+++ src/lib/engn/canvas.rs
@@ -14,6 +14,8 @@ use {
     },
 };
 
+
+/// Stores picture as `Vec<String>` respectively to `charmap` given in the `Conf`
 #[derive(Debug)]
 pub struct Canvas<Lst: AsMaterialList> {
     phantom: PhantomData<Lst>,
@@ -25,6 +27,7 @@ pub struct Canvas<Lst: AsMaterialList> {
 }
 
 impl<Lst: AsMaterialList> Canvas<Lst> {
+    /// Constructs new canvas
     pub fn new(wscr: usize, hscr: usize, charmap: String) -> Self {
         Self {
             phantom: PhantomData,
@@ -36,6 +39,7 @@ impl<Lst: AsMaterialList> Canvas<Lst> {
         }
     }
 
+    /// Updates picture via colliding entities against all camera rays
     pub fn update(&mut self, camera: &Camera, cs: &CoordSys, entities: &Lst) -> ReRes<()> {
         for r in 0..self.hscr {
             self.picture[r] = (0..self.wscr)
@@ -51,6 +55,7 @@ impl<Lst: AsMaterialList> Canvas<Lst> {
         Ok(())
     }
 
+    /// Prints all string to console
     pub fn draw(&self) -> ReRes<()> {
         console::move_cursor(3, 0)?;
         for line in &self.picture {
@@ -59,6 +64,7 @@ impl<Lst: AsMaterialList> Canvas<Lst> {
         Ok(())
     }
 
+    /// Clears all console and shows one message
     pub fn banner(&self, msg: &str, timeout: Duration) -> ReRes<()> {
         let col = self.wscr.saturating_sub(msg.len()) / 2;
         let row = self.hscr / 2;
@@ -71,6 +77,7 @@ impl<Lst: AsMaterialList> Canvas<Lst> {
         Ok(())
     }
 
+    /// Shows message on the second line of console
     pub fn notification(&self, msg: &str) -> ReRes<()> {
         let col = self.wscr.saturating_sub(msg.len()) / 2;
         console::move_cursor(1, col as u16)?;